+++ src/Command/Consume.php
@@ -22,7 +22,7 @@ use Innmind\AMQP\{
 };
 use Innmind\Immutable\{
     Maybe,
-    Either,
+    Attempt,
     Sequence,
     Predicate\Instance,
 };
@@ -48,7 +48,7 @@ final class Consume implements Command
         Channel $channel,
         MessageReader $read,
         State $state,
-    ): Either {
+    ): Attempt {
         $frames = fn(Protocol $protocol): Sequence => $protocol->basic()->consume(
             $channel,
             $this->command,
@@ -57,6 +57,7 @@ final class Consume implements Command
         $sideEffect = match ($this->command->shouldWait()) {
             true => $connection
                 ->request($frames, Method::basicConsumeOk)
+                ->attempt(static fn($failure) => $failure)
                 ->flatMap(fn($frame) => $this->maybeStart(
                     $connection,
                     $channel,
@@ -66,11 +67,12 @@ final class Consume implements Command
                 )),
             false => $connection
                 ->send($frames)
-                ->map(static fn() => $state),
+                ->map(static fn() => $state)
+                ->attempt(static fn($failure) => $failure),
         };
 
-        return $sideEffect->leftMap(
-            fn() => Failure::toConsume($this->command),
+        return $sideEffect->recover(
+            fn() => Attempt::error(Failure::toConsume($this->command)),
         );
     }
 
@@ -93,7 +95,7 @@ final class Consume implements Command
     }
 
     /**
-     * @return Either<Failure, State>
+     * @return Attempt<State>
      */
     private function maybeStart(
         Connection $connection,
@@ -101,14 +103,14 @@ final class Consume implements Command
         MessageReader $read,
         Frame $frame,
         State $state,
-    ): Either {
+    ): Attempt {
         return $frame
             ->values()
             ->first()
             ->keep(Instance::of(Value\ShortString::class))
             ->map(static fn($value) => $value->original()->toString())
             ->either()
-            ->leftMap(fn() => Failure::toConsume($this->command))
+            ->attempt(fn() => Failure::toConsume($this->command))
             ->flatMap(fn($consumerTag) => $this->start(
                 $connection,
                 $channel,
@@ -119,7 +121,7 @@ final class Consume implements Command
     }
 
     /**
-     * @return Either<Failure, State>
+     * @return Attempt<State>
      */
     private function start(
         Connection $connection,
@@ -127,9 +129,9 @@ final class Consume implements Command
         MessageReader $read,
         State $state,
         string $consumerTag,
-    ): Either {
-        /** @var Either<Failure, State|Canceled> */
-        $consumed = Either::right($state);
+    ): Attempt {
+        /** @var Attempt<State|Canceled> */
+        $consumed = Attempt::result($state);
         // here the best approach would be to use recursion to avoid unwrapping
         // the monads but it would end up with a too deep call stack for inifite
         // consumers as each new message would mean a new function call in the
@@ -160,7 +162,7 @@ final class Consume implements Command
     }
 
     /**
-     * @return Either<Failure, State|Canceled>
+     * @return Attempt<State|Canceled>
      */
     private function waitDeliver(
         Connection $connection,
@@ -168,28 +170,30 @@ final class Consume implements Command
         State $state,
         string $consumerTag,
         MessageReader $read,
-    ): Either {
-        /** @var Either<Failure, State|Canceled> */
+    ): Attempt {
         return $connection
             ->wait(Method::basicDeliver)
+            ->attempt(static fn($failure) => $failure)
             ->flatMap(
-                fn($received) => $read($connection)->flatMap(
-                    fn($message) => $this->maybeConsume(
-                        $connection,
-                        $channel,
-                        $read,
-                        $state,
-                        $consumerTag,
-                        $received->frame(),
-                        $message,
+                fn($received) => $read($connection)
+                    ->attempt(static fn($failure) => $failure)
+                    ->flatMap(
+                        fn($message) => $this->maybeConsume(
+                            $connection,
+                            $channel,
+                            $read,
+                            $state,
+                            $consumerTag,
+                            $received->frame(),
+                            $message,
+                        ),
                     ),
-                ),
             )
-            ->leftMap(fn() => Failure::toConsume($this->command));
+            ->recover(fn() => Attempt::error(Failure::toConsume($this->command)));
     }
 
     /**
-     * @return Either<Failure, State|Canceled>
+     * @return Attempt<State|Canceled>
      */
     private function maybeConsume(
         Connection $connection,
@@ -199,7 +203,7 @@ final class Consume implements Command
         string $consumerTag,
         Frame $frame,
         Message $message,
-    ): Either {
+    ): Attempt {
         $destinationConsumerTag = $frame
             ->values()
             ->first()
@@ -232,8 +236,7 @@ final class Consume implements Command
             ->flatMap(static fn($details) => $destinationConsumerTag->map(
                 static fn() => $details, // this manipulation is to make sure the consumerTag is indeed for this consumer
             ))
-            ->either()
-            ->leftMap(fn() => Failure::toConsume($this->command))
+            ->attempt(fn() => Failure::toConsume($this->command))
             ->flatMap(fn($details) => $this->consume(
                 $connection,
                 $channel,
@@ -246,7 +249,7 @@ final class Consume implements Command
     }
 
     /**
-     * @return Either<Failure, State|Canceled>
+     * @return Attempt<State|Canceled>
      */
     private function consume(
         Connection $connection,
@@ -256,7 +259,7 @@ final class Consume implements Command
         Details $details,
         Message $message,
         string $consumerTag,
-    ): Either {
+    ): Attempt {
         return ($this->consume)(
             $state->unwrap(),
             $message,
@@ -270,6 +273,7 @@ final class Consume implements Command
                 $read,
                 $details->deliveryTag(),
                 $consumerTag,
-            );
+            )
+            ->attempt(static fn($failure) => $failure);
     }
 }