+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `Innmind\AMQP\Model\Basic\Message` expiration is now expressed via `Innmind\TimeContinuum\Period`
 - `Innmind\AMQP\Failure` is now an exception that wraps each possible failure object
 - `Innmind\AMQP\Client::run()` now returns an `Innmind\Immutable\Attempt`
+- `Innmind\AMQP\Command::__invoke()` now must return an `Innmind\Immutable\Attempt`
 
 ### Fixed